pyproject.toml
[project]
name = "smolmodels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets",
    "evaluate",
    "wandb",
    "fire",
    "tensorboard",
    "huggingface_hub",
    "pydantic==2.6.3",
    "tabulate==0.9.0",
    "python-dotenv==1.0.1",
    "fsspec==2023.10.0",
    "matplotlib",
    "openai",
    "unidecode",
    "anthropic",
    "pillow==10.4.0",
    "loguru",
    "webdataset",
    "wheel",
    "pytest",
    "kaggle",
    "datacommons",
    "datacommons-pandas>=0.0.3",
    "ipykernel>=6.29.5",
    "rich>=13.9.2",
    "polars>=1.12.0",
    "wrapt-timeout-decorator>=1.5.1",
    "lightning>=2.4.0",
    "ninja>=1.11.1.1",
    "packaging>=24.1",
    "fastapi>=0.115.5",
    "uvicorn>=0.32.1",
    "gradio",
    "scipy",
    "pyright>=1.1.390",
    "google-genai>=0.2.0",
    # pytorch deps
    "transformers==4.47.1",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "peft>=0.13.2",
    "trl>=0.12.0",
    "sentencepiece",
    "einops",
    # add vllm if needed for evals
    "flash-attn>=2.7.0.post2; sys_platform == 'linux'",
    "bitsandbytes>=0.44.1; sys_platform == 'linux'",
    "mistune>=3.1.0",
    "ruff>=0.8.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform != 'darwin'" },
]
torchvision = { index = "pytorch-cu121", marker = "sys_platform != 'darwin'" }
torchaudio = { index = "pytorch-cu121", marker = "sys_platform != 'darwin'" }


[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
